pyproject.toml
[tool.poetry]
name = "foot-stat-scraper"
version = "0.1.0"
description = ""
authors = ["MaksimPerepeliuk <maksim.perepelyuk@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.24.0"
beautifulsoup4 = "^4.9.1"
lxml = "^4.5.2"
psycopg2-binary = "^2.8.5"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
